server/requirements.txt
flask
tensorflow-macos
numpy
pillow
